remove data related to flights, moved to separate class FlightDataManager
